--- webui/react/src/hooks/useModal/Workspace/useModalWorkspaceAddMember.module.scss
@@ -1,14 +1,0 @@
-.base {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  margin-top: 40px;
-
-  .label {
-    color: var(--theme-float-on);
-    margin-bottom: -8px;
-  }
-  :global(p) {
-    margin: 0;
-  }
-}
