--- webui/react/src/hooks/useModal/Workspace/useModalWorkspaceRemoveMember.module.scss
@@ -1,13 +1,0 @@
-.base {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  .label {
-    color: var(--theme-float-on);
-    margin-bottom: -8px;
-  }
-  :global(p) {
-    margin: 0;
-  }
-}
